refactor: repackage files
